Fix build and runtime error on Ubuntu 22.04.
- Fixed SWITCH in VM loop where label address was wrongly casted.
- Use std::ordered_map as default fall back to ext/hash_map in macOS.

+++ src/VM-Run.cpp
@@ -58,7 +58,7 @@ Object VM::runLoop(Object* code, jmp_buf returnPoint, bool returnTable /* = fals
     pc_ = code;
     for (;;) {
         const Object insn = *pc_++;
-        SWITCH((int)insn.val) {
+        SWITCH((intptr_t)insn.val) {
         CASE(HALT)
         {
             return ac_;
